requirements.txt
amqp==5.3.1
async-timeout==5.0.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozendict==2.4.6
fsspec==2025.3.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.1
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
multitasking==0.0.11
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
pymongo==4.11.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.5.3
six==1.17.0
snscrape==0.7.0.20230622
soupsieve==2.6
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.3
tweepy==4.15.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yfinance==0.2.55
